Add ip_visible to InstanceGroup
The ip_visible attribute was added to the InstanceGroup model. This allows
the option for the local ip address to be hidden from any user group.
Database has been updated, as well as the load and update_yml functions.
All yaml files have also been updated to load correctly. ip_visible is
set to true by default. The ability to change this on the scenario/users tab
has not been implemented yet.